+++ test/unit/plugins/coercible/proc_test.rb
@@ -1,0 +1,25 @@
+require 'test_helper'
+
+class ProcCoercionTest < TestBase
+  class ProcObject
+    include Wardrobe
+    attribute :proc, Proc
+  end
+
+  def test_coercion
+    object = ProcObject.new(
+      proc: Proc.new { 1 }
+    )
+
+    assert_equal Proc, object.proc.class
+    assert_equal 1, object.proc.call
+  end
+
+  def test_error
+    assert_raises Wardrobe::Plugins::Coercible::Refinements::UnsupportedError do
+      ProcObject.new(
+        proc: 1
+      )
+    end
+  end
+end